ServiceHelper: start adding a service enum
NewPipe has a list of services it knows about, which is usually equal
to, but sometimes smaller than the one in NewPipeExtractor.

So far the code is mostly passing around random integer `serviceIds`.
This makes it hard to figure out what we are talking about.
Even worse, stuff like `-1` being used to signify “unset service”
makes the code very brittle in many situations.

So this is the first step replacing every such instance with an
exhaustive enum of well-known services.

~~~

For transferring Intent data, we serialize the enum to a string with
`.name()` and then deserialize it on the other side with
`ServiceId.valueOf()`. This should work fine and lead to good error
messages.