VideoDetailFragment: start putting service & player into helper
The `VideoDetailFragmentPlayer` should hold all
player-and-service-related functionality, so that we don’t have to
check everywhere whether both are set.

Plus it should abstract the `VideoDetailFragment` view as much as
possible from the player so that we can start moving things into
Compose.

For now provides an `ifPlayer` block that will make sure the player is
set inside.