VideoDetailFragment: lateinit binding
We can assume that `onCreateView` is always called before anything
else that uses `binding`, so let’s make it lateinit.

The existing null checks are a bit worrying, but it looks like they
must have been from an early version of the app, because turning the
screen with them removed does not trigger any NPE like the comment
suggests.

In fact, I suspect the original problem was that `onServiceConnected`
was called before `onCreateView` in an earlier version of the
app. The logic behind `startService` is quite convoluted, so I’m not
surprised this has been an issue in the past.