VideoDetailFragment: Convert thumbnail to Compose (partially)
Converts the progress bar and duration timestamps to a compose.

The animation part is left out for now, since the preview is hidden in
most cases anyway and does not really need much animation to update
the progress bar. It could be added back later if need should arise.

The primary material color for our UI seems to be wrong? It’s purple
instead of red. But that can be adjusted later.

I feel like this changes focus behavior for large screens? As in you
can’t really focus the player/thumbnail after you leave it. But it’s
kind of weird that that’s only at “thing” for large screens and not on
mobile screens, probably because only large screens are relevant for
TV use?